+++ packages/block/src/BlockHeaderBytes.ts
@@ -1,0 +1,4 @@
+/**
+ * BlockHeaderBuffer is a Buffer array, except for the Verkle PreState which is an array of prestate arrays.
+ */
+export type BlockHeaderBytes = Uint8Array[]
