+++ packages/block/src/CLIQUE_EXTRA_SEAL.ts
@@ -1,0 +1,3 @@
+// Originally from ethereumjs thsu carries the same license
+// Fixed number of extra-data suffix bytes reserved for signer seal
+export const CLIQUE_EXTRA_SEAL = 65
