+++ packages/block/src/CLIQUE_EXTRA_VANITY.ts
@@ -1,4 +1,3 @@
+// From ethereumjs originally and carries same license
 // Fixed number of extra-data prefix bytes reserved for signer vanity
 export const CLIQUE_EXTRA_VANITY = 32
-// Fixed number of extra-data suffix bytes reserved for signer seal
-export const CLIQUE_EXTRA_SEAL = 65
