+++ packages/block/src/ExecutionWitnessBytes.ts
@@ -1,0 +1,1 @@
+export type ExecutionWitnessBytes = Uint8Array
