+++ packages/block/src/RequestBytes.ts
@@ -1,0 +1,1 @@
+export type RequestsBytes = Uint8Array[]
