+++ packages/block/src/TransactionsBytes.ts
@@ -1,0 +1,4 @@
+/**
+ * TransactionsBytes can be an array of serialized txs for Typed Transactions or an array of Uint8Array Arrays for legacy transactions.
+ */
+export type TransactionsBytes = Uint8Array[][] | Uint8Array[]
