+++ packages/block/src/UncleHeaderBytes.ts
@@ -1,0 +1,1 @@
+export type UncleHeadersBytes = Uint8Array[][]
