+++ packages/block/src/WithdrawalsBytes.ts
@@ -1,0 +1,1 @@
+export type WithdrawalsBytes = Uint8Array[]
