+++ packages/block/src/index.ts
@@ -1,7 +1,6 @@
 export { Block } from './block.js'
 export { type BeaconPayloadJson, executionPayloadFromBeaconPayload } from './from-beacon-payload.js'
 export { BlockHeader } from './header.js'
-export { getDifficulty, valuesArrayToHeaderData } from './helpers.js'
 export { blockFromRpc } from './from-rpc.js'
-export { ClRequest } from './ClRequest.js'
-export * from './types.js'
+export type { ClRequest } from './ClRequest.js'
+export * from './WithdrawalsBytes.js'
